refactor : ♻️  유저 정보 서버 통신 로직 분리
